added zod validation for deleting a task

+++ src/useCases/Tasks/DeleteTask/DeleteTaskUseCase.ts
@@ -4,6 +4,12 @@ export class DeleteTaskUseCase {
     constructor(private sqlLiteTaskRepository: SqlLiteTaskRepository) { }
 
     async execute(id: string) {
+        const isValid = await this.sqlLiteTaskRepository.taskExists(id)
+
+        if (!isValid) {
+            throw new Error("Task does not exist")
+        }
+
         await this.sqlLiteTaskRepository.deleteTask(id)
     }
 }
